requirements.txt
google-genai
python-dotenv
docxtpl
pydantic
